demos/requirement.txt
numpy
opencv-python
pillow

